Patch: Metrics testing and unit/regression split (#191)
* add pytest-check and test placeholder for checking results

* add simulation fixture

* add first pass of cost checks

* add missing mooring costs from equipment labor cost breakdown

* remove duplicated port costs from component_costs

* update pandas checks to use pytest-check

* enable split between unit and regression testing

* split tests between unit and regression

* udpate changelog

* add note about split unit and regression options

* use almost_equal to remove floating point tail errors
