+++ CHANGELOG.md
@@ -22,6 +22,12 @@
 - Improved cable, subassembly, and servicing equipment error handling to show which of
   the cables, substations, turbines, or vessels produced the intialization error for
   easier input debugging.
+- Basic tests are now included for limited set of the `Metrics` class with a focus
+  cost summaries.
+- Tests are now roughly split between unit tests and regression tests, where regression
+  tests focus on event timing checks and results checking. Users can now run
+  `pytest --unit` or `pytest --regression` if a subset of the tests are needing to be
+  run with `pytest` still running the entire test suite.
 
 ## v0.10.4 (12 May 2025)
