+++ wombat/core/post_processor.py
@@ -1411,6 +1411,8 @@ class Metrics:
             "transferring crew",
             "traveling",
             "towing",
+            "unmooring",
+            "mooring reconnection",
         ]
         equipment = self.events[self.events[self._equipment_cost] > 0].agent.unique()
         costs = (
@@ -1453,6 +1455,7 @@ class Metrics:
         )
         costs.loc[costs.action == "traveling", "display_reason"] = "Site Travel"
         costs.loc[costs.action == "towing", "display_reason"] = "Towing"
+        costs.loc[costs.action.str.contains("mooring"), "display_reason"] = "Connection"
         costs.loc[costs.action == "mobilization", "display_reason"] = "Mobilization"
         costs.loc[costs.additional.isin(weather_hours), "display_reason"] = (
             "Weather Delay"
@@ -1769,6 +1772,7 @@ class Metrics:
                 & self.events.request_id.str.startswith(("RPR", "MNT")),
                 ["agent", "reason", "request_id", "system_name"],
             ]
+            .drop_duplicates(subset=["request_id"], keep="first")
             .groupby(["agent", "reason", "request_id"])
             .count()
             .reset_index(level=["agent", "reason"], drop=False)
